requirements.txt
oci>=2.100.0
pandas>=1.3.0
PyYAML>=6.0
oracledb>=1.3.0
numpy>=1.21.0